+++ hazelcast/src/main/java/com/hazelcast/internal/tpc/Eventloop.java
@@ -17,13 +17,13 @@
 package com.hazelcast.internal.tpc;
 
 
-import com.hazelcast.internal.tpc.iobuffer.IOBuffer;
+import com.hazelcast.internal.tpc.iobuffer.deprecated.IOBufferImpl;
+import com.hazelcast.internal.tpc.util.CircularQueue;
 import com.hazelcast.internal.util.ThreadAffinity;
 import com.hazelcast.internal.util.ThreadAffinityHelper;
 import com.hazelcast.internal.util.executor.HazelcastManagedThread;
 import com.hazelcast.logging.ILogger;
 import com.hazelcast.logging.Logger;
-import com.hazelcast.internal.tpc.util.CircularQueue;
 import org.jctools.queues.MpmcArrayQueue;
 
 import java.io.Closeable;
@@ -44,10 +44,13 @@ import java.util.concurrent.atomic.AtomicReferenceFieldUpdater;
 import java.util.function.Supplier;
 
 import static com.hazelcast.internal.nio.IOUtil.closeResources;
+import static com.hazelcast.internal.tpc.Eventloop.State.NEW;
+import static com.hazelcast.internal.tpc.Eventloop.State.RUNNING;
+import static com.hazelcast.internal.tpc.Eventloop.State.SHUTDOWN;
+import static com.hazelcast.internal.tpc.Eventloop.State.TERMINATED;
+import static com.hazelcast.internal.tpc.util.Util.epochNanos;
 import static com.hazelcast.internal.util.Preconditions.checkNotNull;
 import static com.hazelcast.internal.util.Preconditions.checkPositive;
-import static com.hazelcast.internal.tpc.Eventloop.State.*;
-import static com.hazelcast.internal.tpc.util.Util.epochNanos;
 import static java.lang.System.getProperty;
 import static java.util.concurrent.atomic.AtomicReferenceFieldUpdater.newUpdater;
 
@@ -346,13 +349,13 @@ public abstract class Eventloop implements Executor {
     }
 
     /**
-     * Offers an {@link IOBuffer} to be processed by this {@link Eventloop}.
+     * Offers an {@link IOBufferImpl} to be processed by this {@link Eventloop}.
      *
-     * @param buff the {@link IOBuffer} to process.
+     * @param buff the {@link IOBufferImpl} to process.
      * @return true if the buffer was accepted, false otherwise.
      * @throws NullPointerException if buff is null.
      */
-    public final boolean offer(IOBuffer buff) {
+    public final boolean offer(IOBufferImpl buff) {
         //todo: Don't want to add localRunQueue optimization like the offer(Runnable)?
 
         if (concurrentRunQueue.offer(buff)) {
@@ -410,8 +413,8 @@ public abstract class Eventloop implements Executor {
                 } catch (Exception e) {
                     logger.warning(e);
                 }
-            } else if (task instanceof IOBuffer) {
-                scheduler.schedule((IOBuffer) task);
+            } else if (task instanceof IOBufferImpl) {
+                scheduler.schedule((IOBufferImpl) task);
             } else {
                 throw new RuntimeException("Unrecognized type:" + task.getClass());
             }
@@ -609,7 +612,7 @@ public abstract class Eventloop implements Executor {
 
         boolean tick();
 
-        void schedule(IOBuffer task);
+        void schedule(IOBufferImpl task);
     }
 
     /**
