+++ hazelcast/src/test/java/com/hazelcast/internal/tpc/iobuffer/IOBufferImplTest.java
@@ -1,6 +1,6 @@
 package com.hazelcast.internal.tpc.iobuffer;
 
-import com.hazelcast.internal.tpc.iobuffer.IOBuffer;
+import com.hazelcast.internal.tpc.iobuffer.deprecated.IOBufferImpl;
 import com.hazelcast.test.HazelcastParallelClassRunner;
 import com.hazelcast.test.annotation.ParallelJVMTest;
 import com.hazelcast.test.annotation.QuickTest;
@@ -8,19 +8,17 @@ import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 
-import static com.hazelcast.internal.nio.Bits.*;
+import static com.hazelcast.internal.nio.Bits.BYTES_INT;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
 
 @RunWith(HazelcastParallelClassRunner.class)
 @Category({QuickTest.class, ParallelJVMTest.class})
-public class IOBufferTest {
+public class IOBufferImplTest {
 
 
     @Test
     public void test() {
-        IOBuffer buf = new IOBuffer(10);
+        IOBufferImpl buf = new IOBufferImpl(10);
 
         int items = 1000;
